% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generator.R
\name{dataset.normal.X1}
\alias{dataset.normal.X1}
\title{simulate data}
\usage{
dataset.normal.X1(family, N, M, theta, response_indicator, seed)
}
\arguments{
\item{family}{"normal" for continuous response}

\item{N}{sample size}

\item{M}{dimension of response}

\item{theta}{vector of true parameter values}

\item{response_indicator}{a vector of integers from 1 up to M, indicating which block the response belongs to}

\item{seed}{set random seed for data generation}
}
\value{
returns a simulated dataset with subject id, response, and one covariate X1
}
\description{
Simulate data for illustration of DIMM
}
\details{
Simulates an artificial dataset with one covariate and response with block AR-1 covariance structure
}
\keyword{data}
\keyword{simulator}
