% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-funcs-package.R
\name{objective.dimm}
\alias{objective.dimm}
\title{goodness-of-fit test for DIMM method}
\usage{
objective.dimm(object, ...)
}
\arguments{
\item{object}{an object of class "dimm"}

\item{...}{additional arguments to be passed}
}
\value{
An object of class "objective.dimm", which contains the value of objective function at the combined estimator, degrees of freedom, and p-value for goodness-of-fit test
}
\description{
Computes a goodness-of-fit test statistics, degrees of freedom, and p-value for given dimm object
}
\details{
Evaluates the objective function at the combined estimator
}
